backend/portfolio/portfolio_backend/requirements.txt
Django==5.2.4
psycopg[binary]==3.2.9
drf-auth-kit[social]==0.3.4
django-allauth[socialaccount]==65.10.0
dj-rest-auth==5.0.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
pillow==11.3.0
firebase-admin==7.0.0
